requirements.txt
asgiref==3.2.7
backcall==0.1.0
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
Django==3.0.5
django-channels==0.7.0
django-debug-toolbar==2.2
django-extensions==2.2.9
django-filter==2.2.0
django-tables2==2.3.1
django-webpack-loader==0.7.0
factory-boy==2.12.0
Faker==4.0.2
idna==2.9
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
oauthlib==3.1.0
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pydot==1.4.1
pydotplus==2.0.2
Pygments==2.6.1
pygraphviz==1.5
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.9
Werkzeug==1.0.1
